Convert coordinates to long for bit operations
Ensure integer coordinates before using bitwise operations by converting gx and gy to long tensors. This prevents potential errors from floating point types when performing bit shifts or masks.

Relates to I130-6

+++ src/vit_pytorch/layers/embeddings/fractal_path.py
@@ -217,7 +217,11 @@ class VectorizedPathEncoder:
         # 缩放: grid_x = cx * grid_size // img_size
         gx = cx * grid_size // safe_img_size  # [B, N]
         gy = cy * grid_size // safe_img_size  # [B, N]
-        
+
+        # I130-6: 转换为整数类型以支持位运算
+        gx = gx.long()
+        gy = gy.long()
+
         # 确保在有效范围内
         gx = gx.clamp(0, grid_size - 1)
         gy = gy.clamp(0, grid_size - 1)
